renaming and moving of functions
